Removed unused imports

+++ src/main/java/com/endava/config/ConfigApplication.java
@@ -3,11 +3,11 @@ package com.endava.config;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.MalformedURLException;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
 
+import org.apache.commons.io.FileUtils;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,6 +22,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.core.MongoTemplate;
 
 import com.endava.config.annotation.EnableMongoConfigServer;
+import com.endava.config.repository.DevRepository;
+import com.endava.config.repository.ProdRepository;
+import com.endava.config.repository.QARepository;
 
 import lombok.extern.log4j.Log4j2;
 import springfox.documentation.builders.PathSelectors;
@@ -41,6 +44,15 @@ public class ConfigApplication implements CommandLineRunner {
 	@Autowired
 	private MongoTemplate mongoTemplate;
 
+	@Autowired
+	private DevRepository devRepository;
+
+	@Autowired
+	private QARepository qaRepository;
+
+	@Autowired
+	private ProdRepository prodRepository;
+
 	public static void main(String[] args) {
 		SpringApplication.run(ConfigApplication.class, args);
 	}
@@ -52,13 +64,35 @@ public class ConfigApplication implements CommandLineRunner {
 
 	private void initializeConfigs() throws IOException, ParseException, URISyntaxException {
 		for (Profile profile : Profile.values()) {
-				log.debug(String.format("Attempting to insert configs for service %s", profile));
+			int collectionSize = -1;
+			switch (profile) {
+			case DEV:
+				collectionSize = devRepository.findAll().size();
+				break;
+
+			case QA:
+				collectionSize = qaRepository.findAll().size();
+				break;
+
+			case PROD:
+				collectionSize = prodRepository.findAll().size();
+				break;
+
+			default:
+				log.info("DEFAULTED");
+				break;
+			}
+
+			if (collectionSize == 0) {
+				log.debug("%s is empty!", profile);
+				log.debug(String.format("Attempting to insert configuration to profile %s", profile));
 				FileReader reader = new FileReader(getFileFromResources(getConfigFilePath(profile)));
 				JSONParser jsonParser = new JSONParser();
 				String JSONString = jsonParser.parse(reader).toString();
 				log.debug(String.format("Inserting configurations %s", JSONString));
 				mongoTemplate.dropCollection(profile.toString().toLowerCase());
 				mongoTemplate.save(JSONString, profile.toString().toLowerCase());
+			}
 		}
 	}
 
@@ -67,16 +101,18 @@ public class ConfigApplication implements CommandLineRunner {
 	}
 
 	private File getFileFromResources(String filename) throws URISyntaxException, MalformedURLException, IOException {
-		ClassLoader classLoader = getClass().getClassLoader();
 
-		URL resource = classLoader.getResource(filename);
+//		 this is the path within the jar file
+		InputStream input = ConfigApplication.class.getResourceAsStream(filename);
+		if (input == null) {
+//	     this is how we load file within editor
+			input = ConfigApplication.class.getClassLoader().getResourceAsStream(filename);
+		}
 
-		if (resource == null) {
-			throw new IllegalArgumentException("File not found!");
-		} else {
-			URI uri = resource.toURI();
-			File file = new File(uri.toURL().getPath());
-			return file;
+		try (InputStream is = input) {
+			File tempFile = new File(filename);
+			FileUtils.copyInputStreamToFile(is, tempFile);
+			return tempFile;
 		}
 	}
